requirements.txt
django>=2.0.7
django-ckeditor
django-mathfilters
Pillow
